Some more street stuff.
Plans to add barriers around map edge that are unpassable. Plans to build a forest area that bars the player to a forced map follow. Plans to add a small city in the forest. Plans to add a bit more to the city and then some prop work to everything. Map size will be cut either in half, or a quarter cut off of it.